+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
-## [Unreleased]
+## [7.5.5] - 2024-04-17
 
 ### Added
 
@@ -1412,7 +1412,9 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/7.5.3...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/7.5.5...HEAD
+[7.5.5]: https://github.com/dipdup-io/dipdup/compare/7.5.4...7.5.5
+[7.5.4]: https://github.com/dipdup-io/dipdup/compare/7.5.3...7.5.4
 [7.5.3]: https://github.com/dipdup-io/dipdup/compare/7.5.2...7.5.3
 [7.5.2]: https://github.com/dipdup-io/dipdup/compare/7.5.1...7.5.2
 [7.5.1]: https://github.com/dipdup-io/dipdup/compare/7.5.0...7.5.1